requirements-dev.txt
pytest==5.3.0
pytest-django==3.7.0
pytest-factoryboy==2.0.3